Moved the CPU data model by pydantic to a new file
